move executor to top level, add dataframe execute support
