+++ mars/dataframe/execution/datasource.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 # Copyright 1999-2018 Alibaba Group Holding Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,3 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from ..expressions.datasource.dataframe import DataFrameDataSource
+
+
+def _dataframe_pandas_data_source(ctx, chunk):
+    ctx[chunk.key] = chunk.op.data
+
+
+def register_data_source_handler():
+    from ...executor import register
+
+    register(DataFrameDataSource, _dataframe_pandas_data_source)
