+++ mars/tensor/execution/tests/test_core_execute.py
@@ -16,7 +16,7 @@
 
 import unittest
 
-from mars.tensor.execution.core import Executor
+from mars.executor import Executor
 from mars.tensor import ones
 from mars.session import LocalSession, Session
